requirements.txt
Django==4.2.2
pandas==1.3.5
gunicorn==20.1.0
crispy-bootstrap5==0.7
django-crispy-forms==2.0
requests==2.31.0
seaborn==0.12.2
